futureQuant/requirements.txt
# futureQuant 依赖列表

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0  # Parquet支持

# 数据获取
akshare>=1.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 配置管理
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0

# 机器学习
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# 深度学习（可选）
# torch>=2.0.0

# 优化
optuna>=3.3.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# 爬虫（可选，用于动态页面）
# playwright>=1.40.0
# drissionpage>=4.0.0

# 其他工具
tqdm>=4.65.0
python-dateutil>=2.8.0
